+++ src/app/search/SearchBar.tsx
@@ -130,40 +130,55 @@ const Row = React.memo(
     isPhonePortrait: boolean;
     isTabAutocompleteItem: boolean;
     onClick(e: React.MouseEvent, item: SearchItem): void;
-  }) => (
-    <>
-      <AppIcon className={styles.menuItemIcon} icon={searchItemIcons[item.type]} />
-      <span className={styles.menuItemQuery}>
-        {item.type === SearchItemType.Help ? (
-          t('Header.FilterHelpMenuItem')
-        ) : item.highlightRange ? (
-          <HighlightedText
-            text={item.query}
-            startIndex={item.highlightRange[0]}
-            endIndex={item.highlightRange[1]}
-            className={styles.textHighlight}
-          />
-        ) : (
-          item.query
+  }) => {
+    function highlight(text: string, section: string) {
+      return item.highlightRange?.section === section ? (
+        <HighlightedText
+          text={text}
+          startIndex={item.highlightRange.range[0]}
+          endIndex={item.highlightRange.range[1]}
+          className={styles.textHighlight}
+        />
+      ) : (
+        text
+      );
+    }
+
+    return (
+      <>
+        <AppIcon className={styles.menuItemIcon} icon={searchItemIcons[item.type]} />
+        <p className={styles.menuItemQuery}>
+          {item.query.header && highlight(item.query.header, 'header')}
+          {item.type === SearchItemType.Help ? (
+            t('Header.FilterHelpMenuItem')
+          ) : (
+            <span
+              className={clsx({
+                [styles.namedQueryBody]: item.query.header !== undefined,
+              })}
+            >
+              {highlight(item.query.body, 'body')}
+            </span>
+          )}
+        </p>
+        <span className={styles.menuItemHelp} />
+        {!isPhonePortrait && isTabAutocompleteItem && (
+          <KeyHelp className={styles.keyHelp} combo="tab" />
         )}
-      </span>
-      <span className={styles.menuItemHelp} />
-      {!isPhonePortrait && isTabAutocompleteItem && (
-        <KeyHelp className={styles.keyHelp} combo="tab" />
-      )}
-      {!isPhonePortrait && highlighted && <KeyHelp className={styles.keyHelp} combo="enter" />}
-      {(item.type === SearchItemType.Recent || item.type === SearchItemType.Saved) && (
-        <button
-          type="button"
-          className={styles.deleteIcon}
-          onClick={(e) => onClick(e, item)}
-          title={t('Header.DeleteSearch')}
-        >
-          <AppIcon icon={closeIcon} />
-        </button>
-      )}
-    </>
-  )
+        {!isPhonePortrait && highlighted && <KeyHelp className={styles.keyHelp} combo="enter" />}
+        {(item.type === SearchItemType.Recent || item.type === SearchItemType.Saved) && (
+          <button
+            type="button"
+            className={styles.deleteIcon}
+            onClick={(e) => onClick(e, item)}
+            title={t('Header.DeleteSearch')}
+          >
+            <AppIcon icon={closeIcon} />
+          </button>
+        )}
+      </>
+    );
+  }
 );
 
 // TODO: break filter autocomplete into its own object/helpers... with tests
@@ -267,7 +282,7 @@ function SearchBar(
     stateReducer,
     initialIsOpen: isPhonePortrait && mainSearchBar,
     defaultHighlightedIndex: liveQuery ? 0 : -1,
-    itemToString: (i) => i?.query || '',
+    itemToString: (i) => i?.query.fullText || '',
     onInputValueChange: ({ inputValue, type }) => {
       setLiveQuery(inputValue || '');
       debouncedUpdateQuery(inputValue || '');
@@ -324,7 +339,7 @@ function SearchBar(
   const deleteSearch = useCallback(
     (e: React.MouseEvent, item: SearchItem) => {
       e.stopPropagation();
-      dispatch(searchDeleted(item.query));
+      dispatch(searchDeleted(item.query.fullText));
     },
     [dispatch]
   );
@@ -357,11 +372,11 @@ function SearchBar(
   const tabAutocompleteItem =
     highlightedIndex > 0 && items[highlightedIndex]?.type === SearchItemType.Autocomplete
       ? items[highlightedIndex]
-      : items.find((s) => s.type === SearchItemType.Autocomplete && s.query !== liveQuery);
+      : items.find((s) => s.type === SearchItemType.Autocomplete && s.query.fullText !== liveQuery);
   const onKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
     if (e.key === 'Tab' && !e.altKey && !e.ctrlKey && tabAutocompleteItem && isOpen) {
       e.preventDefault();
-      setInputValue(tabAutocompleteItem.query);
+      setInputValue(tabAutocompleteItem.query.fullText);
       if (tabAutocompleteItem.highlightRange) {
         selectionRef.current = tabAutocompleteItem.highlightRange[1];
       }
@@ -377,7 +392,7 @@ function SearchBar(
       items[highlightedIndex]?.type === SearchItemType.Recent
     ) {
       e.preventDefault();
-      dispatch(searchDeleted(items[highlightedIndex].query));
+      dispatch(searchDeleted(items[highlightedIndex].query.fullText));
     } else if (e.key === 'Enter' && !isOpen && liveQuery) {
       // Show search results on "Enter" with a closed menu
       dispatch(toggleSearchResults());
@@ -393,7 +408,7 @@ function SearchBar(
               className={clsx(styles.menuItem, {
                 [styles.highlightedItem]: highlightedIndex === index,
               })}
-              key={`${item.type}${item.query}`}
+              key={`${item.type}${item.query.fullText}`}
               {...getItemProps({ item, index })}
             >
               <Row